// client/src/api/model/incidental.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */
import type { DiceModifier } from './diceModifier';
import type { HealEffect } from './healEffect';
import type { ResultsModifier } from './resultsModifier';

export interface Incidental {
  /** Dice added to or removed from rolls when this incidental is activated */
  diceModifiers: DiceModifier[];
  /** Fixed result symbols added to rolls when this incidental is activated */
  resultsModifiers: ResultsModifier[];
  /** Healing applied to wounds or strain when this incidental is activated */
  healEffects: HealEffect[];
}