// client/src/api/model/healSource.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Genesys API Documentation
 * Interactive API documentation
 * OpenAPI spec version: 1.0
 */

export type HealSource = typeof HealSource[keyof typeof HealSource];


export const HealSource = {
  Fixed: 'Fixed',
  Cybernetics: 'Cybernetics',
  Characteristic: 'Characteristic',
} as const;